Keep moving stuff into 'ingests', and tidy up the namespace
